Refactor block uniqueness normalization pass into separate file.
