+++ src/cwe_checker_lib/src/intermediate_representation/blk.rs
@@ -40,22 +40,6 @@ pub struct Blk {
 }
 
 impl Term<Blk> {
-    /// Return a clone of `self` where the given suffix is appended to
-    /// the TIDs of all contained terms (the block itself and all `Jmp`s and `Def`s).
-    ///
-    /// Note that all TIDs of jump targets (direct, indirect and return targets) are left unchanged.
-    fn clone_with_tid_suffix(&self, suffix: &str) -> Self {
-        let mut cloned_block = self.clone();
-        cloned_block.tid = cloned_block.tid.with_id_suffix(suffix);
-        for def in cloned_block.term.defs.iter_mut() {
-            def.tid = def.tid.clone().with_id_suffix(suffix);
-        }
-        for jmp in cloned_block.term.jmps.iter_mut() {
-            jmp.tid = jmp.tid.clone().with_id_suffix(suffix);
-        }
-        cloned_block
-    }
-
     /// Remove indirect jump target addresses for which no corresponding target block exists.
     /// Return an error message for each removed address.
     pub fn remove_nonexisting_indirect_jump_targets(
